hdbrag_dspy_optimized/pyproject.toml
[project]
name = "DSPyer-web"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "dspy-ai==2.6.27",
    "langchain-openai>=0.3.35",
    "langgraph>=0.6.11",
    "llama-index>=0.14.10",
    "openai>=2.13.0",
    "python-dotenv>=1.2.1",
]
